+++ src/main/java/com/eclipsekingdom/playerplot/util/storage/CallbackQuery.java
@@ -1,7 +1,0 @@
-package com.eclipsekingdom.playerplot.util.storage;
-
-public interface CallbackQuery<T> {
-
-    void onQueryDone(T t);
-
-}