+++ src/main/java/me/sword7/playerplot/plotdeed/ILoot.java
@@ -1,0 +1,21 @@
+package me.sword7.playerplot.plotdeed;
+
+import org.bukkit.inventory.ItemStack;
+
+import java.util.List;
+
+public interface ILoot {
+
+    ItemStack valueOf(String s) throws Exception;
+
+    String getListLabel();
+
+    List<String> getLootTypes();
+
+    String getRoot();
+
+    String getType();
+
+    ItemStack getDefault();
+
+}