+++ src/main/java/me/sword7/playerplot/user/UserData.java
@@ -1,0 +1,27 @@
+package me.sword7.playerplot.user;
+
+public class UserData {
+
+    private int unlockedPlots;
+
+    public UserData() {
+        this.unlockedPlots = 0;
+    }
+
+    public UserData(int unlockedPlots) {
+        this.unlockedPlots = unlockedPlots;
+    }
+
+    public int getUnlockedPlots() {
+        return unlockedPlots;
+    }
+
+    public void unlockPlot() {
+        unlockedPlots++;
+    }
+
+    public void lockPlots(int amount) {
+        unlockedPlots -= amount;
+    }
+
+}