+++ src/main/java/me/sword7/playerplot/util/storage/CallbackQuery.java
@@ -1,0 +1,7 @@
+package me.sword7.playerplot.util.storage;
+
+public interface CallbackQuery<T> {
+
+    void onQueryDone(T t);
+
+}